requirements.txt
edge-tts
moviepy
Pillow
anthropic
